Ensure that all possible node states are registered with micrometer.
Metrics systems do not all deal well with metrics that are only *sometimes* present.
For example, in prometheus it is very annoying if a certain metric is appearing and disappearing,
because it makes it hard to create decent graphs of it and to create alerts on it.
A metric that has not appeared for some time is regarded as non-existent to prometheus.