Disable/enable health fixer plugin when performing fixer integration test
Otherwise, we have the change that the fixer already indexes the nodes immediately after we have purged them. Then our test would fail because we never detect that the nodes have been purged in solr.

Allow for more time for solr indexing/purging to complete, it seems that
sometimes it takes more than 60 seconds.