Make SolrIndexValidationHealthProcessorPluginTest clearer when they fail
Instead of using Stream#allMatch() to check that all items have a certain health status, collect them into a set,
and then compare the set to a singleton set with the expected value.
When the assertion fails, that has a clearer error message about what statuses were actually returned, instead of just having a fail message.