Add reporter to log immediately when unhealthy node is detected
The reason for adding this reporter is twofold:
1. When an unhealthy node is detected, you want to know it as soon as
possible, and not waiting until the health processor cycle has completed
(which can take days on a large repository)
2. With the limitation on the number of stored reports, we may not be
able to get all generated reports at the end of the cycle. Reporting
them in a streaming fashion allows users to search their logs for all
unhealthy nodes, albeit in a less handy way than with a summary.