Log that reports have been dropped at the end of a cycle
Write a log message about number of reports that have been dropped at
the end of the cycle instead of at the moment that the dropping starts.

It is more helpful to log at the end, because we can then report how
many reports have been dropped. It is also closer to the summary
reported by SummaryLoggingHealthReporter