Log number of dropped reports in SummaryLoggingHealthReporter
This is to make sure that administrators don't misunderstand the list of
unhealthy nodes logged by the SummaryLoggingHealthReporter as complete
when there are reports that have been dropped.

Calling reports "dropped" may be confusing with regards to whether
fixers have been run on those reports or not.

Do not touch the header, but add a note at the end of the list of
reports that there are additional, not logged, messages.