Place webscript into main context
Webscripts should be part of the main context, so the bean is
not recreated when the subsystem is restarted.

If the webscript bean is recreated; it would be uninitialized, which
results in problems with the webscript registry and the admin console
(e.g. NullPointerException when trying to access the webscript)